Search radius
Search radius now works as intended. Edge cases are also intuitive to users.

+++ lib/data_files/gua_globals.dart
@@ -2,7 +2,7 @@ double gpsLatitude = -1;
 double gpsLongitude = -1;
 
 bool openNow = false;
-double searchRadius = -1;
+int searchRadius = -1;
 String sortBy = "best_match";
 String sortByUI = "Best Match";
 String cuisine = "";
